% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsdata.R
\docType{data}
\name{cdsdata}
\alias{cdsdata}
\title{CDS quotes from market}
\format{
An object of class \code{"data.frame"}.
}
\source{
Thomson Reuters, CDS quotes of Unicredit on 2017-01-23
}
\usage{
data(cdsdata)
}
\description{
\itemize{
\item \code{Maturity}: Maturities of cds contracts expressed in years;
\item \code{Par.Spread}: CDS rates quotes, spread that nullify the present value of the two legs;
\item \code{ED.Zero.Curve}: EURIBOR interest rates (risk-free)
}
}
\keyword{datasets}
