% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_categorical_vars.R
\name{detect_categorical_vars}
\alias{detect_categorical_vars}
\title{Detect Categorical Variables}
\usage{
detect_categorical_vars(dataset, n_unique = 20)
}
\arguments{
\item{dataset}{A data frame containing the input data.}

\item{n_unique}{An integer specifying the maximum number of unique values a column can have to be considered categorical. Default: 20.}
}
\value{
A character vector containing the names of the columns classified as categorical.
}
\description{
A function that identifies columns in a dataset that can be considered categorical based on the number of unique values and their data type.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
