% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_input.R
\name{na_input}
\alias{na_input}
\title{Handle Missing Values in Dataset}
\usage{
na_input(dataset, target, n_unique = 20)
}
\arguments{
\item{dataset}{A data frame containing the input data to be processed.}

\item{target}{A character string indicating the name of the target variable in the dataset. This column will not be altered during this function.}

\item{n_unique}{An integer specifying the maximum number of unique values a column can have to be considered categorical, Default: 20}
}
\value{
A data frame with missing values filled.
}
\description{
A function that handles missing values in a dataset by filling them with appropriate values based on the data type. Numerical columns are filled with the mean, while categorical columns are filled with the mode.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}
 \code{\link[logger]{log_level}}
}
