% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoder.R
\name{one_hot_encoder}
\alias{one_hot_encoder}
\title{One-Hot Encoding for Categorical Variables}
\usage{
one_hot_encoder(dataset, categorical_vars)
}
\arguments{
\item{dataset}{A data frame containing the input data to be encoded.}

\item{categorical_vars}{A character vector specifying the names of the categorical variables to be one-hot encoded.}
}
\value{
A data frame containing the original dataset (excluding the original categorical variables) combined with the newly created dummy variables for each level of the specified categorical variables.
}
\description{
A function that performs one-hot encoding on specified categorical variables in a dataset, creating dummy variables for each unique level of the categorical variables.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[caret]{dummyVars}}
}
