% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\title{Preprocess Dataset for Analysis}
\usage{
pre_process(dataset, target, id_var, na_threshold = 0.5)
}
\arguments{
\item{dataset}{A data frame containing the input data to be preprocessed}

\item{target}{A character string indicating the name of the target variable #' in the dataset.}

\item{id_var}{A character string indicating the column to be treated as #' identifier. These columns are excluded during preprocessing.}

\item{na_threshold}{A numeric value between 0 and 1 indicating the maximum proportion of missing values allowed in a column. Default: 0.5.}
}
\value{
A preprocessed data frame with cleaned and optimized features.
}
\description{
A function that performs several preprocessing steps on a
dataset, such as removing unwanted columns, handling missing values,
and addressing multicollinearity and making the data ready for analysis
or modeling.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[logger]{log_level}}
}
