% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unwanted_cols.R
\name{remove_unwanted_cols}
\alias{remove_unwanted_cols}
\title{Remove Unwanted Columns}
\usage{
remove_unwanted_cols(
  dataset,
  target,
  na_threshold = 0.5,
  remove_col_NA = FALSE,
  remove_unique_cols = FALSE,
  remove_duplicate_cols = FALSE,
  remove_ld_cols = FALSE,
  remove_cor_cols = FALSE
)
}
\arguments{
\item{dataset}{A data frame containing the input data to be processed.}

\item{target}{A character string specifying the name of the target variable.}

\item{na_threshold}{A numeric value between 0 and 1 indicating the maximum proportion of missing values allowed in a column. Default: 0.5.}

\item{remove_col_NA}{A logical value indicating whether to remove columns with a proportion of missing values exceeding `na_threshold`. Default: FALSE.}

\item{remove_unique_cols}{A logical value indicating whether to remove columns with only unique values. Default: FALSE.}

\item{remove_duplicate_cols}{A logical value indicating whether to remove duplicate columns. Default: FALSE.}

\item{remove_ld_cols}{A logical value indicating whether to remove columns with linear dependencies. Default: FALSE.}

\item{remove_cor_cols}{A logical value indicating whether to remove columns with high correlations (correlation coefficient > 0.8). Default: FALSE.}
}
\value{
A data frame with the unwanted columns removed,
}
\description{
A function that removes unwanted columns from a dataset based on user-specified criteria, such as high missing values, unique values, duplicates, linear dependencies, or high correlations.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{bind_cols}}
 \code{\link[logger]{log_level}}
 \code{\link[caret]{findLinearCombos}}
}
