% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modeling.R
\name{run_modeling}
\alias{run_modeling}
\title{Wrapper function to run the package}
\usage{
run_modeling(
  dataset,
  target,
  id_var,
  test_size = 0.2,
  maxdepth = 30,
  na_threshold = 0.5
)
}
\arguments{
\item{dataset}{A data frame containing the input data.}

\item{target}{A character string indicating the name of the target variable in the dataset.}

\item{id_var}{A character string indicating the column to be treated as #' identifier. These columns are excluded during preprocessing.}

\item{test_size}{The fraction of the dataset to be used as the test set. Default: 0.2.}

\item{maxdepth}{The maximum depth of the decision tree. Default: 30.}

\item{na_threshold}{PARAM_DESCRIPTION, Default: 0.5}

\item{na_thresholdA}{numeric value between 0 and 1 indicating the maximum proportion of missing values allowed in a column. Default: 0.5.}
}
\value{
Run pre_process and modeling and return a list containing the trained model and performance metrics
}
\description{
A function that performs pre_process and modeling
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
